Optimize AoC 25 day 2 Python solution
Instead of checking any number of partitions where
the target digit number modulo partition size is 0,
only the prime numbers of the digit numbers are taken
to be the partition sizes -> which reduces the number
of checks by alot. The execution was reduced from 5s
to 3s. (I'm sure i will not understand this commit
msg if I'll read it in the future)